refactor: move & simplify legacy code
