+++ go.mod
@@ -2,4 +2,7 @@ module github.com/join-com/protoc-gen-ts
 
 go 1.16
 
-require github.com/golang/protobuf v1.4.3
+require (
+	github.com/golang/protobuf v1.4.3
+	google.golang.org/protobuf v1.25.0
+)
