+++ legacy/base/generator.go
