+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/join-com/protoc-gen-ts/generator"
+	"github.com/join-com/protoc-gen-ts/legacy/generator"
 )
 
 func main() {
